Empleado CRUD v1
## Added EmpleadoController
All of the endpoints are ready to test, waiting for changes.
##Updated: EmpleadoEntity, EmpleadoRepository and EmpleadoService
Some of the methods in Repo and Service were changed or removed

+++ src/main/java/mx/tecnm/hris/empleados/EmpleadoService.java
@@ -3,10 +3,9 @@ package mx.tecnm.hris.empleados;
 import mx.tecnm.hris.empleados.enums.Estado;
 import mx.tecnm.hris.empleados.enums.NivelAcademico;
 import mx.tecnm.hris.empleados.enums.Sexo;
-import mx.tecnm.hris.empleados.enums.TipoContrato;
+import mx.tecnm.hris.empleados.enums.TipoMovimiento;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -24,10 +23,6 @@ public class EmpleadoService {
         return empleadoRepository.findById(id);
     }
 
-    public Optional<EmpleadoEntity> findByNumeroTarjeta(Long id) {
-        return findById(id);
-    }
-
     public Optional<EmpleadoEntity> findByRfc(String rfc) {
         return empleadoRepository.findByRfc(rfc);
     }
@@ -36,7 +31,7 @@ public class EmpleadoService {
         return empleadoRepository.findByCurp(curp);
     }
 
-    public List<EmpleadoEntity> findAll(){
+    public List<EmpleadoEntity> findAll() {
         return empleadoRepository.findAll();
     }
 
@@ -60,44 +55,36 @@ public class EmpleadoService {
         return empleadoRepository.findBySexo(sexo).stream().collect(Collectors.toList());
     }
 
-    public List<EmpleadoEntity> findByTipoContrato(TipoContrato tipoContrato) {
-        return empleadoRepository.findByTipoContrato(tipoContrato).stream().collect(Collectors.toList());
+    public List<EmpleadoEntity> findByTipoContrato(TipoMovimiento tipoMovimiento) {
+        return empleadoRepository.findByTipoContrato(tipoMovimiento).stream().collect(Collectors.toList());
     }
 
-    public void updateStatusByRfc(String rfc, Estado estado) {
-        empleadoRepository.actualizarEstadoPorRfc(rfc, estado);
+    public int updateEstadoByRfc(String rfc, Estado estado) {
+        return empleadoRepository.actualizarEstadoPorRfc(rfc, estado);
     }
 
-    public void updateStatusByCurp(String curp, Estado estado) {
-        empleadoRepository.actualizarEstadoPorCurp(curp, estado);
+    public int updateEstadoByCurp(String curp, Estado estado) {
+        return empleadoRepository.actualizarEstadoPorCurp(curp, estado);
     }
 
-    public void updateStatusById(Long id, Estado estado) {
-        empleadoRepository.actualizarEstadoPorId(id, estado);
-    }
-
-    public List<EmpleadoEntity> findByNombreYSexo(String nombre, Sexo sexo) {
-        return empleadoRepository.findByNombreAndSexo(nombre, sexo).stream().collect(Collectors.toList());
+    public int updateEstadoById(Long id, Estado estado) {
+        return empleadoRepository.actualizarEstadoPorId(id, estado);
     }
 
     public List<EmpleadoEntity> findByNivelAcademicoMayorIgual(NivelAcademico nivelAcademico) {
         return empleadoRepository.findByNivelAcademicoGreaterThanEqual(nivelAcademico).stream().collect(Collectors.toList());
     }
 
-    public List<EmpleadoEntity> findByFechaIngresoBeforeAndEstadoEquals(LocalDate fecha, Estado estado) {
-        return empleadoRepository.findByFechaIngresoBeforeAndEstadoEquals(fecha, estado).stream().collect(Collectors.toList());
-    }
-
     public EmpleadoEntity save(EmpleadoEntity empleado) {
         return empleadoRepository.save(empleado);
     }
 
-    public List<EmpleadoEntity> saveAll(List<EmpleadoEntity> empleados) {
-        return empleadoRepository.saveAll(empleados);
-    }
-
-    public void deleteById(Long id) {
-        empleadoRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (empleadoRepository.existsById(id)) {
+            empleadoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
